Adiciona cadastro e exclusão de status do item

+++ src/main/java/br/com/fatec/mogi/inventory_service/coreService/service/impl/CategoriaItemServiceImpl.java
@@ -3,11 +3,14 @@ package br.com.fatec.mogi.inventory_service.coreService.service.impl;
 import br.com.fatec.mogi.inventory_service.common.web.response.CustomPageResponseDTO;
 import br.com.fatec.mogi.inventory_service.coreService.domain.exception.CategoriaJaCadastradaException;
 import br.com.fatec.mogi.inventory_service.coreService.domain.exception.CategoriaNaoEncontradaException;
+import br.com.fatec.mogi.inventory_service.coreService.domain.exception.CategoriaNaoPodeSerExcluidaException;
 import br.com.fatec.mogi.inventory_service.coreService.domain.model.CategoriaItem;
 import br.com.fatec.mogi.inventory_service.coreService.repository.CategoriaItemRepository;
+import br.com.fatec.mogi.inventory_service.coreService.repository.ItemRepository;
 import br.com.fatec.mogi.inventory_service.coreService.service.CategoriaItemService;
 import br.com.fatec.mogi.inventory_service.coreService.web.request.AtualizarCategoriaItemRequestDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.request.CadastrarCategoriaItemRequestDTO;
+import br.com.fatec.mogi.inventory_service.coreService.web.request.ConsultarCategoriaItemRequestDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.response.BuscarCategoriasResponseDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.response.CategoriaItemResponseDTO;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +22,8 @@ import org.springframework.stereotype.Service;
 public class CategoriaItemServiceImpl implements CategoriaItemService {
 
 	private final CategoriaItemRepository categoriaItemRepository;
+
+	private final ItemRepository itemRepository;
 
 	@Override
 	public BuscarCategoriasResponseDTO buscar() {
@@ -37,6 +42,10 @@ public class CategoriaItemServiceImpl implements CategoriaItemService {
 	@Override
 	public void deletar(Long id) {
 		categoriaItemRepository.findById(id).orElseThrow(CategoriaNaoEncontradaException::new);
+		var categoriaVinculada = itemRepository.existsByCategoriaItemId(id);
+		if (categoriaVinculada) {
+			throw new CategoriaNaoPodeSerExcluidaException();
+		}
 		categoriaItemRepository.deleteById(id);
 	}
 
@@ -48,8 +57,9 @@ public class CategoriaItemServiceImpl implements CategoriaItemService {
 	}
 
 	@Override
-	public CustomPageResponseDTO<CategoriaItemResponseDTO> buscarPaginado(Pageable pageable) {
-		var categorias = categoriaItemRepository.findPaginado(pageable);
+	public CustomPageResponseDTO<CategoriaItemResponseDTO> buscarPaginado(ConsultarCategoriaItemRequestDTO dto,
+			Pageable pageable) {
+		var categorias = categoriaItemRepository.findPaginado(dto, pageable);
 		return CustomPageResponseDTO.<CategoriaItemResponseDTO>builder()
 			.content(categorias.getContent())
 			.page(categorias.getNumber())
