+++ src/main/java/br/com/fatec/mogi/inventory_service/coreService/service/impl/LocalizacaoServiceImpl.java
@@ -2,12 +2,15 @@ package br.com.fatec.mogi.inventory_service.coreService.service.impl;
 
 import br.com.fatec.mogi.inventory_service.common.web.response.CustomPageResponseDTO;
 import br.com.fatec.mogi.inventory_service.coreService.domain.exception.LocalizacaoNaoEncontradaException;
+import br.com.fatec.mogi.inventory_service.coreService.domain.exception.LocalizacaoNaoPodeSerExcluidaException;
 import br.com.fatec.mogi.inventory_service.coreService.domain.mapper.LocalizacaoMapper;
 import br.com.fatec.mogi.inventory_service.coreService.domain.model.Localizacao;
+import br.com.fatec.mogi.inventory_service.coreService.repository.ItemRepository;
 import br.com.fatec.mogi.inventory_service.coreService.repository.LocalizacaoRepository;
 import br.com.fatec.mogi.inventory_service.coreService.service.LocalizacaoService;
 import br.com.fatec.mogi.inventory_service.coreService.web.request.AtualizarLocalizacaoRequestDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.request.CadastrarLocalizacaoRequestDTO;
+import br.com.fatec.mogi.inventory_service.coreService.web.request.ConsultarLocalizacaoRequestDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.response.BuscarLocalizacaoResponseDTO;
 import br.com.fatec.mogi.inventory_service.coreService.web.response.LocalizacaoResponseDTO;
 import jakarta.transaction.Transactional;
@@ -20,6 +23,8 @@ import org.springframework.stereotype.Service;
 public class LocalizacaoServiceImpl implements LocalizacaoService {
 
 	private final LocalizacaoRepository localizacaoRepository;
+
+	private final ItemRepository itemRepository;
 
 	private final LocalizacaoMapper localizacaoMapper;
 
@@ -35,8 +40,9 @@ public class LocalizacaoServiceImpl implements LocalizacaoService {
 	}
 
 	@Override
-	public CustomPageResponseDTO<LocalizacaoResponseDTO> buscarPaginado(Pageable pageable) {
-		var localizacoes = localizacaoRepository.findPaginado(pageable);
+	public CustomPageResponseDTO<LocalizacaoResponseDTO> buscarPaginado(ConsultarLocalizacaoRequestDTO dto,
+			Pageable pageable) {
+		var localizacoes = localizacaoRepository.findPaginado(dto, pageable);
 		return CustomPageResponseDTO.<LocalizacaoResponseDTO>builder()
 			.page(localizacoes.getNumber())
 			.size(localizacoes.getSize())
@@ -57,6 +63,10 @@ public class LocalizacaoServiceImpl implements LocalizacaoService {
 	@Override
 	public void deletar(Long id) {
 		localizacaoRepository.findById(id).orElseThrow(LocalizacaoNaoEncontradaException::new);
+		var localizacaoVinculada = itemRepository.existsByLocalizacaoId(id);
+		if (localizacaoVinculada) {
+			throw new LocalizacaoNaoPodeSerExcluidaException();
+		}
 		localizacaoRepository.deleteById(id);
 	}
